Create README.md
The objective of this task is to detect hate speech in tweets. For the sake of simplicity, we say a tweet contains hate speech if it has a racist or sexist sentiment associated with it. So, the task is to classify racist or sexist tweets from other tweets.
XGBoost with Word2Vec model has given us the best performance. My F1-score on public leaderboard is 0.7375.